+++ src/gui/globalHotKeys/X11ErrorLogger.h
@@ -1,8 +1,8 @@
 /*
- * Copyright (C) 2019 Damir Porobic <https://github.com/damirporobic>
+ * Copyright (C) 2020 Damir Porobic <damir.porobic@gmx.com>
  *
  * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
+ * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2 of the License, or
  * (at your option) any later version.
  *
@@ -11,29 +11,28 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
- * You should have received a copy of the GNU Lesser General Public License
+ * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor,
  * Boston, MA 02110-1301, USA.
  */
 
-#include "KeyHandlerFactory.h"
+#ifndef KSNIP_X11ERRORLOGGER_H
+#define KSNIP_X11ERRORLOGGER_H
 
-AbstractKeyHandler* KeyHandlerFactory::create()
+#include <QString>
+
+#include <X11/Xlib.h>
+
+class X11ErrorLogger
 {
-#if defined(__APPLE__)
-    return new DummyKeyHandler;
-#endif
+public:
+	X11ErrorLogger();
+	~X11ErrorLogger() = default;
 
-#if defined(__linux__)
-    if(PlatformChecker::instance()->isWayland()) {
-	    return new DummyKeyHandler;
-    } else {
-	    return new X11KeyHandler;
-    }
-#endif
+private:
 
-#if  defined(_WIN32)
-    return new WinKeyHandler;
-#endif
-}
+	static int errorHandler(Display *d, XErrorEvent *e);
+};
+
+#endif //KSNIP_X11ERRORLOGGER_H