+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed: Logout canceled by 'ksnip' under KDE. ([#281](https://github.com/ksnip/kImageAnnotator/issues/281))
 * Fixed: Ksnip not displayed on the monitor (off screen). ([#307](https://github.com/ksnip/kImageAnnotator/issues/307))
 * Fixed: CTRL+Q to quit Ksnip not working. ([#308](https://github.com/ksnip/kImageAnnotator/issues/308))
+* Fixed: Global Hotkeys not working with activated Num and Caps Lock under X11. ([#310](https://github.com/ksnip/kImageAnnotator/issues/310))
 
 ## Release 1.6.1
 * Change: Allow opening link directly to image without opening in browser. ([#248](https://github.com/ksnip/kImageAnnotator/issues/248))